Move IcCheckingInputs into forms folder
